pkg/systimemon: a system time monitor
