+++ server/util.go
@@ -398,22 +398,3 @@ func diffRegionKeyInfo(origin *RegionInfo, other *RegionInfo) string {
 
 	return strings.Join(ret, ",")
 }
-
-// StartTimeMonitor will call systimeErrHandler if system time jump backward.
-func StartTimeMonitor(now func() time.Time, systimeErrHandler func() bool) {
-	log.Info("start system time monitor")
-	tick := time.NewTicker(100 * time.Millisecond)
-	defer tick.Stop()
-	for {
-		last := now()
-		select {
-		case <-tick.C:
-			if now().Sub(last) < 0 {
-				log.Errorf("system time jump backward, last:%v", last)
-				if !systimeErrHandler() {
-					break
-				}
-			}
-		}
-	}
-}
