+++ server/util_test.go
@@ -14,7 +14,6 @@
 package server
 
 import (
-	"sync/atomic"
 	"time"
 
 	. "github.com/pingcap/check"
@@ -40,29 +39,4 @@ func (s *testMinMaxSuite) TestMinDuration(c *C) {
 	c.Assert(minDuration(time.Minute, time.Second), Equals, time.Second)
 	c.Assert(minDuration(time.Second, time.Minute), Equals, time.Second)
 	c.Assert(minDuration(time.Second, time.Second), Equals, time.Second)
-}
-
-var _ = Suite(&testSystimeMonitorSuite{})
-
-type testSystimeMonitorSuite struct{}
-
-func (s *testSystimeMonitorSuite) TestSystimeMonitor(c *C) {
-	jumpForward := new(int32)
-	atomic.StoreInt32(jumpForward, 0)
-	trigged := false
-	go StartTimeMonitor(
-		func() time.Time {
-			if !trigged {
-				trigged = true
-				return time.Now()
-			}
-
-			return time.Now().Truncate(2 * time.Second)
-		}, func() bool {
-			atomic.AddInt32(jumpForward, 1)
-			return false
-		})
-
-	time.Sleep(1 * time.Second)
-	c.Assert(atomic.LoadInt32(jumpForward), Equals, int32(1))
 }
